// angular/base-bff-api/projects/bff-api/src/lib/models/userRoleGroup.ts
export class UserRoleGroup {
    maxNoOfUser: number;
    maxNoOfUserCreated: number;
    userGroupCode: string;
    userRoleGroupCode: string;
    parentRoleGroup: string;
    userRoleGroupDesc: string;
    userGroupDesc: string;
    userType: string;
    userTypeDesc: string;
    cntryCd: string;
    selCountry: boolean;
    selBranch: boolean;
    createByProfId: boolean;
    createByBranchId: boolean;
    systemCreate: boolean;
}